Major refactor of HTTP parser allowing streaming of large request bodies
This refactor simplifies the parser to some extent and simplifies the consuming code as well. It also adds support for streamed bodies that aren't chunked. You can now set a memory limit and the body will be streamed as chunks if the entire request body cannot fit in that memory limit.