Allow to listen on specific interface
The http_server_listen() and http_server_listen_poll() function
listen by default on all available interfaces. In order to be
able to listen only on a specific interface, the functions
http_server_listen_addr() and http_server_listen_addr_poll()
allow to pass the IP address if the interface to listen on as
second parameter.